requirements.txt
Flask==2.1.2
Flask_Authorize==0.2.6
Flask_Dance==6.0.0
Flask_Login==0.6.1
Flask_Security==3.0.0
Flask_SQLAlchemy==2.5.1
Flask_WTF==1.0.1
gunicorn==20.1.0
Jinja2==2.11.3
opencensus==0.9.0
opencensus-ext-azure==1.1.4
opencensus_ext_flask==0.8.0
psycopg2==2.9.3
python-dotenv==0.20.0
Werkzeug==2.1.2
WTForms==3.0.1
